Fixed bug, problem wasn't with allocator, first MB of RAM will always be kept reserved

+++ managers/filesystem/ext2/ext2.h
@@ -8,7 +8,7 @@
 
 uint32_t _EXT2_Initialize(FileDescriptor *desc);
 uint32_t _EXT2_Filesystem_OpenFile(FileDescriptor *desc, const char *filename, int flags, int perms);
-uint8_t _EXT2_Filesystem_ReadFile(FileDescriptor *desc, UID id, uint8_t *buffer, size_t size);
+uint64_t _EXT2_Filesystem_ReadFile(FileDescriptor *desc, UID id, uint8_t *buffer, size_t size);
 uint64_t _EXT2_Filesystem_SeekFile(FileDescriptor *desc, uint32_t fd, uint64_t offset, int whence);
 uint8_t _EXT2_Filesystem_CloseFile(FileDescriptor *desc, uint32_t fd);
 uint8_t _EXT2_Filesystem_DeleteFile(FileDescriptor *desc, const char *file);
